// src/types/theme.ts
export type ThemeVariant = 'primary' | 'secondary' | 'success' | 'danger'| 'error' | 'warning' | 'info' | 'dark' | 'light'
export type ComponentSize = 'default' | 'sm' | 'md' | 'lg' | 'xl'

export enum ThemeColors {
  PRIMARY = '#1967D2',
  SECONDARY = '#C2E7FF',
  SUCCESS = '#198754',
  INFO = '#0dcaf0',
  DANGER = '#1967D2',
  ERROR = '#1967D2',
  WARNING = '#ffc107',
  LIGHT = '#f8f9fa',
  DARK = '#343a40',
}

export type TypeVariant = 'text' | 'flat' | 'elevated' | 'tonal' | 'outlined' | 'plain' | 'filled'